KFSPTS-10559: Code changes from PR feedback
